[Move `@kbn/config-schema` to server] `content_management`
