+++ src/plugins/content_management/common/rpc/constants.ts
@@ -5,9 +5,6 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import { schema } from '@kbn/config-schema';
-import { validateVersion } from '@kbn/object-versioning-utils';
-
 export const procedureNames = [
   'get',
   'bulkGet',
@@ -19,12 +16,3 @@ export const procedureNames = [
 ] as const;
 
 export type ProcedureName = (typeof procedureNames)[number];
-
-export const versionSchema = schema.number({
-  validate: (value) => {
-    const { result } = validateVersion(value);
-    if (!result) {
-      return 'must be an integer';
-    }
-  },
-});
