+++ src/plugins/content_management/common/rpc/create.ts
@@ -5,32 +5,8 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import { schema } from '@kbn/config-schema';
 import type { Version } from '@kbn/object-versioning';
-import { itemResultSchema } from './common';
-import { versionSchema } from './constants';
-
-import type { ItemResult, ProcedureSchemas } from './types';
-
-export const createSchemas: ProcedureSchemas = {
-  in: schema.object(
-    {
-      contentTypeId: schema.string(),
-      version: versionSchema,
-      // --> "data" to create a content will be defined by each content type
-      data: schema.recordOf(schema.string(), schema.any()),
-      options: schema.maybe(schema.object({}, { unknowns: 'allow' })),
-    },
-    { unknowns: 'forbid' }
-  ),
-  out: schema.object(
-    {
-      contentTypeId: schema.string(),
-      result: itemResultSchema,
-    },
-    { unknowns: 'forbid' }
-  ),
-};
+import type { ItemResult } from './types';
 
 export interface CreateIn<
   T extends string = string,
