+++ src/plugins/content_management/common/rpc/delete.ts
@@ -5,35 +5,7 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import { schema } from '@kbn/config-schema';
 import type { Version } from '@kbn/object-versioning';
-import { versionSchema } from './constants';
-
-import type { ProcedureSchemas } from './types';
-
-export const deleteSchemas: ProcedureSchemas = {
-  in: schema.object(
-    {
-      contentTypeId: schema.string(),
-      id: schema.string({ minLength: 1 }),
-      version: versionSchema,
-      options: schema.maybe(schema.object({}, { unknowns: 'allow' })),
-    },
-    { unknowns: 'forbid' }
-  ),
-  out: schema.object(
-    {
-      contentTypeId: schema.string(),
-      result: schema.object(
-        {
-          success: schema.boolean(),
-        },
-        { unknowns: 'forbid' }
-      ),
-    },
-    { unknowns: 'forbid' }
-  ),
-};
 
 export interface DeleteIn<T extends string = string, Options extends void | object = object> {
   contentTypeId: T;
