+++ src/plugins/content_management/common/rpc/get.ts
@@ -5,33 +5,8 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import { schema } from '@kbn/config-schema';
 import type { Version } from '@kbn/object-versioning';
-import { itemResultSchema } from './common';
-import { versionSchema } from './constants';
-
-import type { ItemResult, ProcedureSchemas } from './types';
-
-export const getResultSchema = schema.object(
-  {
-    contentTypeId: schema.string(),
-    result: itemResultSchema,
-  },
-  { unknowns: 'forbid' }
-);
-
-export const getSchemas: ProcedureSchemas = {
-  in: schema.object(
-    {
-      contentTypeId: schema.string(),
-      id: schema.string({ minLength: 1 }),
-      version: versionSchema,
-      options: schema.maybe(schema.object({}, { unknowns: 'allow' })),
-    },
-    { unknowns: 'forbid' }
-  ),
-  out: getResultSchema,
-};
+import type { ItemResult } from './types';
 
 export interface GetIn<T extends string = string, Options extends void | object = object> {
   id: string;
