+++ src/plugins/content_management/common/rpc/update.ts
@@ -5,33 +5,8 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import { schema } from '@kbn/config-schema';
 import type { Version } from '@kbn/object-versioning';
-import { itemResultSchema } from './common';
-import { versionSchema } from './constants';
-
-import type { ItemResult, ProcedureSchemas } from './types';
-
-export const updateSchemas: ProcedureSchemas = {
-  in: schema.object(
-    {
-      contentTypeId: schema.string(),
-      id: schema.string({ minLength: 1 }),
-      version: versionSchema,
-      // --> "data" to update a content will be defined by each content type
-      data: schema.recordOf(schema.string(), schema.any()),
-      options: schema.maybe(schema.object({}, { unknowns: 'allow' })),
-    },
-    { unknowns: 'forbid' }
-  ),
-  out: schema.object(
-    {
-      contentTypeId: schema.string(),
-      result: itemResultSchema,
-    },
-    { unknowns: 'forbid' }
-  ),
-};
+import type { ItemResult } from './types';
 
 export interface UpdateIn<
   T extends string = string,
