+++ src/plugins/content_management/server/rpc/procedures/search.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { rpcSchemas } from '../../../common/schemas';
+import { rpcSchemas } from '../schemas';
 import type { SearchIn } from '../../../common';
 import { getContentClientFactory } from '../../content_client';
 import type { ProcedureDefinition } from '../rpc_service';
