+++ src/plugins/content_management/server/rpc/schemas/index.ts
@@ -6,6 +6,4 @@
  * Side Public License, v 1.
  */
 
-// exporting schemas separately from the index.ts file to not include @kbn/schema in the public bundle
-// should be only used server-side or in jest tests
 export { schemas as rpcSchemas } from './rpc';
